pyproject.toml
[project]
name = "simple-video-preview"
version = "0.1.0"
description = "Minimal macOS camera preview window with CLI and JSON config."
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
  "pyobjc-framework-AVFoundation>=11.1",
  "pyobjc-framework-Cocoa>=11.1",
]

[project.scripts]
video-preview = "simple_video_preview.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/simple_video_preview"]
